desafio9-mocks_y_normalizr
1) useNullAsDefault: true al usar SQLite3.
2) Generada la vista de 5 productos aleatorios usando faker, implemantada en la ruta api/productos-test a través del método GET.

+++ desafio9-mocks_y_normalizr/db/db.js
@@ -10,7 +10,7 @@ class DB {
     static async addProduct(newProduct) {
         const data = {
             title: newProduct.title,
-            price: parseInt(newProduct.price),
+            price: parseFloat(newProduct.price),
             thumbnail: newProduct.thumbnail
         }
         await insertOnTable(database.sql, 'products', data);
